add tests for batch-exchange with relation target
